backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pgvector==0.3.6
voyageai==0.3.2
anthropic==0.40.0
feedparser==6.0.11
trafilatura==1.12.2
beautifulsoup4==4.12.3
requests==2.32.3
httpx==0.27.2
python-multipart==0.0.12
slack-bolt==1.21.0
rank-bm25==0.2.2
python-dotenv==1.0.1
pydantic-settings==2.6.1
tenacity==9.0.0
python-dateutil==2.9.0
